move optional funcs from views.py to utils.py
